Renames module file
