+++ src/domain/checklists/checklistActions.ts
@@ -1,6 +1,6 @@
 import { Checklist } from './models';
 import { API_ROUTES } from '../../api/apiRoutes';
-import { sendApiRequest } from '../../api/types';
+import { sendApiRequest } from '../../api/module';
 
 
 export const fetchChecklists = (jwt: string, categoryId: number) => sendApiRequest<undefined, Checklist[]>(
