+++ src/domain/habits/habitActions.ts
@@ -1,6 +1,6 @@
 import { Habit } from './models';
 import { API_ROUTES } from '../../api/apiRoutes';
-import { sendApiRequest } from '../../api/types';
+import { sendApiRequest } from '../../api/module';
 
 
 export const fetchHabits = (jwt: string) => sendApiRequest<undefined, Habit[]>(
