+++ src/domain/tasks/taskActions.ts
@@ -1,6 +1,6 @@
 import { Task } from './models';
 import { API_ROUTES } from '../../api/apiRoutes';
-import { sendApiRequest } from '../../api/types';
+import { sendApiRequest } from '../../api/module';
 
 
 export const fetchTasks = (jwt: string, checklistId: number) => sendApiRequest<undefined, Task[]>(
